+++ sqlparser/lib/sqlparser.dart
@@ -4,6 +4,7 @@ library sqlparser;
 export 'src/analysis/analysis.dart';
 export 'src/ast/ast.dart';
 export 'src/engine/module/fts5.dart' show Fts5Extension;
+export 'src/engine/module/json1.dart' show Json1Extension;
 export 'src/engine/module/module.dart';
 export 'src/engine/options.dart';
 export 'src/engine/sql_engine.dart';